requirements.txt
streamlit==1.25.0
plotly==5.15.0
pandas==2.0.3
numpy==1.25.2
folium==0.14.0
matplotlib==3.7.2
matplotlib-inline==0.1.6
haversine==2.8.0
streamlit-folium==0.7.0
Pillow==10.0.0


